Update flake.nix: changed URL literals to strings
URL literals cause confusion as they use the same x:x syntax as lambdas. The use of these is considered to be a bad practice. Apart from that, Lix which is a Nix fork dropped support of this feature with its 2.92 release making building of Lumen with it result in an error.